Fix for dev profvis
For complicated reasons, we had to change the evaluation strategy in profvis and it no longer modifies the environment that it's run in. However, I think `bench::mark()` probably better suited for your use case anyway, especially since it comes with a convenient way to check the outputs are equal.